pyproject.toml
[tool.poetry]
name = "it-solutions"
version = "0.1.0"
description = ""
authors = ["alex <sashason1990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^72.1.0"
drf-spectacular = "^0.27.2"
coreapi = "^2.3.3"
psycopg2-binary = "^2.9.9"
pytest-django = "^4.8.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
